// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  SELLER: 'SELLER',
  STOCK_MANAGER: 'STOCK_MANAGER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const StockMovementType = {
  PURCHASE: 'PURCHASE',
  SALE: 'SALE',
  SALE_CANCELLATION: 'SALE_CANCELLATION',
  ADJUSTMENT: 'ADJUSTMENT',
  INITIAL: 'INITIAL'
} as const

export type StockMovementType = (typeof StockMovementType)[keyof typeof StockMovementType]


export const SaleStatus = {
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type SaleStatus = (typeof SaleStatus)[keyof typeof SaleStatus]


export const PaymentMethod = {
  CASH: 'CASH',
  CARD: 'CARD',
  MOBILE_MONEY: 'MOBILE_MONEY',
  CREDIT: 'CREDIT',
  OTHER: 'OTHER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PurchaseStatus = {
  RECEIVED: 'RECEIVED',
  CANCELLED: 'CANCELLED'
} as const

export type PurchaseStatus = (typeof PurchaseStatus)[keyof typeof PurchaseStatus]